Move tool bar to below
Users are having a hard time finding the
features to add text. Hopefully this will make
it easier for them to identify which buttons
they can click on.

Turning scrolling back on for this page, as I think the issue I was having
was specific to the duck duck go browser